x/ml/types/message_language_test.go
package types

import (
	"testing"

	"github.com/blockentropy/entropy/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgLanguage_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgLanguage
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgLanguage{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgLanguage{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
